adding board context
