+++ src/App.css
@@ -1,6 +1,7 @@
 .container {
   display: flex;
   justify-content: center;
+  margin-top: 2rem;
 }
 
 .container-row {
