--- src/AppAdmin.tsx
@@ -1,5 +1,0 @@
-import BoardAdmin from "./components/BoardAdmin";
-
-export default function AppAdmin() {
-  return <BoardAdmin />;
-}
